refactor: move models to folders
