+++ Recademy.Application/Services/Implementations/ReviewResponseService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Recademy.Application.Services.Abstractions;
 using Recademy.Core.Models;
+using Recademy.Core.Models.Reviews;
 using Recademy.Core.Types;
 using Recademy.DataAccess;
 using Recademy.Shared.Dtos.Reviews;
