+++ Recademy.Core/Models/Skills/UserSkill.cs
@@ -1,14 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
+using Recademy.Core.Models.Users;
 
-namespace Recademy.Core.Models
+namespace Recademy.Core.Models.Skills
 {
     public class UserSkill
     {
-        [ForeignKey("User")] 
+        [ForeignKey("User")]
         public int UserId { get; set; }
         public User User { get; set; }
 
-        [ForeignKey("Skill")] 
+        [ForeignKey("Skill")]
         public string SkillName { get; set; }
         public Skill Skill { get; set; }
     }
