+++ Recademy.DataAccess/Repositories/IReviewRepository.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
-using Recademy.Core.Models;
+using Recademy.Core.Models.Projects;
+using Recademy.Core.Models.Reviews;
 using Recademy.Core.Types;
 
 namespace Recademy.DataAccess.Repositories
