+++ Recademy.DataAccess/Repositories/Implementations/ReviewRepository.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
-using Recademy.Core.Models;
+using Recademy.Core.Models.Projects;
+using Recademy.Core.Models.Reviews;
 using Recademy.Core.Types;
 
 namespace Recademy.DataAccess.Repositories.Implementations
