+++ Recademy.DataAccess/Repositories/Implementations/UserRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 using Microsoft.EntityFrameworkCore;
-using Recademy.Core.Models;
+using Recademy.Core.Models.Users;
 using Recademy.Core.Types;
 
 namespace Recademy.DataAccess.Repositories.Implementations
