+++ Recademy.Mock/Mocker.cs
@@ -4,7 +4,10 @@ using System.Configuration;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Recademy.Api;
-using Recademy.Core.Models;
+using Recademy.Core.Models.Projects;
+using Recademy.Core.Models.Reviews;
+using Recademy.Core.Models.Skills;
+using Recademy.Core.Models.Users;
 using Recademy.Core.Types;
 using Recademy.DataAccess;
 using Recademy.Mock.Extensions;
