+++ Recademy.Shared/Dtos/Reviews/ReviewResponseInfoDto.cs
@@ -1,4 +1,4 @@
-﻿using Recademy.Core.Models;
+﻿using Recademy.Core.Models.Reviews;
 using Recademy.Core.Tools;
 using Recademy.Core.Types;
