+++ Recademy.Test/MockerTest.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
 using Recademy.Api;
-using Recademy.Core.Models;
+using Recademy.Core.Models.Users;
 using Recademy.DataAccess;
 using Recademy.Mock;
 using Recademy.Test.Tools;
